fix(web): merge session-rescope shim into existing proxy
Next 16 already has src/proxy.ts (middleware successor); a separate
middleware.ts alongside it fails the build.
